requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
openai==1.99.8
pinecone-client==6.0.0
python-dotenv==1.1.1
pydantic==2.11.7
python-docx==1.2.0
requests==2.32.4
tqdm==4.67.1
